relação serviços estágio/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-marshmallow==1.3.0
marshmallow-sqlalchemy==0.26.1
python-dotenv==1.0.1